#17 restored edited summary view files from earlier commit
